move program.rs into own module
